refactor: 역할이 명확하지 않은 Cars 객체에 자동차 경주에 참여하는 자동차 목록을 생성하는 책임 부여
