+++ docs/SETP3-README.md
@@ -41,7 +41,7 @@
 * [x] `kotlin.io`의 `readLine()`을 이용한 값 입력 처리
 
 ## 코드 리뷰 피드백 내용 정리
-* [ ] 역할이 명확하지 않은 Cars 객체에 자동차 경주에 참여하는 자동차 목록을 생성하는 책임 부여
+* [x] 역할이 명확하지 않은 Cars 객체에 자동차 경주에 참여하는 자동차 목록을 생성하는 책임 부여
 * [ ] 불변성을 가져야 하는 가변 타입으로 선언된 필드를 불변 객체로 변경 
 * [ ] repeat()를 이용하여 for문으로 작성된 반복문의 가독성 개선
