+++ src/main/kotlin/step3/racingcar/domain/Cars.kt
@@ -1,5 +1,7 @@
 package step3.racingcar.domain
 
-class Cars(val elements: MutableList<Car> = mutableListOf()) {
+class Cars(val count: Int) {
+    private val elements: List<Car> = List(count) { Car() }
+    fun elements() = elements
     fun size() = elements.size
 }
