+++ src/main/kotlin/step3/racingcar/service/RacingCarService.kt
@@ -1,12 +1,11 @@
 package step3.racingcar.service
 
 import step3.racingcar.domain.Cars
-import step3.racingcar.utils.CarGenerator
 import step3.racingcar.utils.RandomNumberGenerator
 
 class RacingCarService {
     fun play(carCount: Int, roundCount: Int): Cars {
-        val cars: Cars = CarGenerator.generate(carCount)
+        val cars = Cars(carCount)
 
         for (i in 1..roundCount) {
             playEachRound(i, cars)
@@ -15,7 +14,7 @@ class RacingCarService {
     }
 
     private fun playEachRound(currentRound: Int, cars: Cars) {
-        cars.elements.forEach {
+        cars.elements().forEach {
             val randomValue = RandomNumberGenerator.generate()
             it.race(currentRound, randomValue)
         }
