+++ src/test/kotlin/step3/racingcar/domain/CarsTest.kt
@@ -1,13 +1,12 @@
-package step3.racingcar.utils
+package step3.racingcar.domain
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
-import step3.racingcar.domain.Cars
 
-internal class CarGeneratorTest : BehaviorSpec({
-    given("CarGenerator") {
+internal class CarsTest : BehaviorSpec({
+    given("Cars") {
         `when`("특정 숫자가 주어지면") {
-            val actual: Cars = CarGenerator.generate(5)
+            val actual = Cars(5)
             then("숫자만큼 자동차가 생성된다.") {
                 actual.size() shouldBe 5
             }
